Fix mismatching range checks in finding a new target when gathering.
While gathering resources entities may be tasked to gather near a position. The range check for that is 10 m. However, the relaxed range check for the walking state is this.DefaultRelaxedMaxRange, which is 12 meter, currently. Meaning it would lead to a slow infinite loop.

Differential revision: D3839
Comments by: @wraitii
Tested by: @gameboy (https://wildfiregames.com/forum/topic/38131-tutorial-bug/)



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@25249 3db68df2-c116-0410-a063-a993310a9797