Check for replay compatibility when opening summary screen
Summary screen and mainly its statistics can be influenced by mod as much as replay itself with mod removing or adding classes and trackers to the game.
Summary screen allows to replay the match without checking compatibility, that is fine I think if it is prevented to see summary of incompatible replay in a first place.

Therefore this is adding compatibility check for opening summary of replay.

Differential revision: D3799
Fixes: #6119
Tested by: @Langbart


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@25268 3db68df2-c116-0410-a063-a993310a9797