De-singletonify CConfigDB, make hooks RAII.
CConfigDB is no longer a singleton, and its members are no longer static.
This makes it possible to have several instances of it, useful for testing of if we ever want to have different config profiles.

Also updates hooks to delete themselves on destruction, to fit a more natural RAII pattern. Long-lived objects must take care to destroy hooks before ConfigDB is reset.

Differential Revision: https://code.wildfiregames.com/D3866



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@25326 3db68df2-c116-0410-a063-a993310a9797