Generalize icon-specific tooltips to a tag, streamline code.
We can currently specify a custom tooltip for an icon. This extends that to a generic tag `[tooltip='something']...[/tooltip]` for any particular piece of text, and removes the iconTooltip hardcoding by instead using a virtual `GetTooltipText` call since all IGUIObject descendants have a tooltip. This is both cleaner & more efficient.
Removes the ability to set a custom tooltip style (that doesn't seem particularly useful), it'll reuse the style of the CText itself.

Differential Revision: https://code.wildfiregames.com/D3895



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@25353 3db68df2-c116-0410-a063-a993310a9797