Enable specifying the trigger action independently of the trigger name.
Trigger functions/actions were tied to the trigger 'identifier'. This is inconvenient if several triggers should call the same function, e.g. for range triggers.

By setting an 'action' in the trigger data, one can now specify the function to call independently of the trigger name.

For clarity, the 'action' parameter in RegisterTrigger is changed to 'name'. By default, the action still matches the trigger name for compatibility and convenience.

Differential Revision: https://code.wildfiregames.com/D3903



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@25374 3db68df2-c116-0410-a063-a993310a9797