Fix auto-complete chat entries for Biome names in GameSetup
The line altered here was returning an Array of Objects. This array would
eventually passed to code that expected it to be an Array of Strings. This was
causing an error to thrown that could be replicated by going to the Multiplayer
GameSetup, entering text (e.g. "Winter", "Dry", "lol") into the chat field, and
pressing `Tab` to trigger the auto-complete.

This line now returns an Array containing the user-facing names of the Biomes
available to the selected map (if it has them), which is what this method was
returning prior to r25077.


Related: phab:rP25077, r25077



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@25403 3db68df2-c116-0410-a063-a993310a9797