Don't hardcode the "0ad" resource into lobby XMPP & hosting
XMPP JID has a concept of 'resources', which can be used to differentiate multiple clients of the same account. 

We currently hardcode this 'resource' to '0ad' in two places:
- The 0 A.D. client always uses '0ad'
- The network code expects a host resource to be '0ad' when connecting.

As noted in https://code.wildfiregames.com/rP21520#31215, it is less effort to store the JI
D directly. This patch does that. It also makes 0 A.D. use a different resource each time.
Note that resources ought not contain particular information, as the XMPP server is free to
 clobber it. I keep '0ad-' here for debug purposes.

This allows:
- multiple 0 A.D. instances to log on the lobby at the same time (not massively useful, but good for debugging sometimes)
- hosting a game with a custom resource, which will potentially make it easier to have dedi
cated servers on one account.

Note that hosting multiple games on one account is currently not supported and will have weird behaviour on the lobbybots side. They should be upgraded independently of this.

Refs #3556

Differential Revision: https://code.wildfiregames.com/D3500



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@25407 3db68df2-c116-0410-a063-a993310a9797