Speed up GetAvailableMods for archives.
Refs rP15677 and rP21823

When a .zip file is encountered by the VFS population, it reads the info for all files in the archives. This is quite slow for the public archive (400-500ms on my computer), which means calling GetEngineInfo() repeatedly is impossible.

By only opening the external mod.json, we skip most of the work. The archive can still be opened if needed as fallback.


Differential Revision: https://code.wildfiregames.com/D3216



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@25446 3db68df2-c116-0410-a063-a993310a9797