Tweak sound distance attenuation & add configurability
Following D3108 / rP24382, sounds are attenuated by their actual distance. However, as noted by players on A24, the dropoff is stark and, when zoomed out, can easily result in not hearing things that are happening in the middle of the screen.

The new default settings almost double the max-range, change the minimum range to have greater dynamic range, and reduce the stereo depth slightly to better match the default camera FOV.

These are stored per sound-grounp, possibly allowing future tweaks on a per-soundgroup basis, and are configurable.

Tested by: Imarok

Discussed over mail with: Samulis, Porru

Differential Revision: https://code.wildfiregames.com/D3612



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@25547 3db68df2-c116-0410-a063-a993310a9797