Fix crash when pushing a page while pushing a page.
Because the page stack is a vector, if during PushPage, a new page is pushed, the vector may re-allocate. This 'pulls the rug out' from underneath the code stack that originally pushed, which then crashes.

To fix this, use a deque, since push/pop won't invalidate references.

Reported by: Imarok.

Based on a patch by: Imarok (tests are his)

Differential Revision: https://code.wildfiregames.com/D4037



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@25616 3db68df2-c116-0410-a063-a993310a9797