Allow mods to say they should be ignored in replay/MP compatibility checks
Since it is very non-trivial to determine which mods change checksums and which don't, this relies on modder goodwill (and on verification on our end for signed mods).

The declaration is an optional "ignoreInCompatibilityChecks" boolean in mod.json

Also rework slightly the MP lobby mod display to always show the host mods in a clear manner.

Differential Revision: https://code.wildfiregames.com/D3968



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@25634 3db68df2-c116-0410-a063-a993310a9797