Remove territory root from fortresses.
Reverts D1762 / rP24393.
Because any territory connected to a root does not decay, this made it possible to anchor CC territories by adding a fortress somewhere, thus making it much, much harder to push against an enemy territory.
This led to a much more turtley gameplay and overall made the game worse.

For now, I do not reintroduce the larger territory range (which was reduced in rP24393).

Differential Revision: https://code.wildfiregames.com/D3989



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@25697 3db68df2-c116-0410-a063-a993310a9797