Lower Forge technology costs, introduce a 3rd level.
Follows rP24604 / rP24725.
Forge technologies are cheaper in A24, but still too expensive to use directly, making the forge not immediately useful.
This changes the technologies from 2 levels to 3, and further decreases costs.
Instead of 600/400 then 900/600, costs are 200/100, 350/250 then 500/400. This makes the Town Phase techs immediately accessible, and lengthens impact of City Phase with stronger late techs.

Iberians and Mauryas had a special "wootz steel" tech that gave swordsmen +20% damage, this is folded into the 3rd level. I have not changed the effect because the tech was cheap, but we shall see.

Patch by: Nescio

Comments by: borg-, Valihrant

Differential Revision: https://code.wildfiregames.com/D3679



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@25726 3db68df2-c116-0410-a063-a993310a9797