Prevent OOM crash in Reference Tree on error
The structTree, in case of errors, could have enough items to draw to trigger an OOM failure in the Arena allocator.

This fixes that by hiding elements by default and some c++ memory optimisation (mostly, this should make all platforms take the same memory footprint for VisibleObject).

Discussed with vladislavbelov and s0600204

Differential Revision: https://code.wildfiregames.com/D4114



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@25746 3db68df2-c116-0410-a063-a993310a9797