UnitMotion hack to fix units being stuck near corners.
I think there is a small mismatch between CheckMovement & the vertex pathfinder when computing passability because one is ray-based and the other just uses edges. However, it's out of my reach to fix it for now.
This can lead to units being stuck near building edges occasionally. This introduces an un-intrusive recovery strategy (aka a hack) in HandleObstructedMove that should get the units unstuck.

Should fix #6114 (to a sufficient extent anyways)

Differential Revision: https://code.wildfiregames.com/D4162



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@25786 3db68df2-c116-0410-a063-a993310a9797