Fix serialisation crash when patrolling.
r24360 / rP24360 introduced waiting at the patrol points, but it was possible that at such a point an enemy was noticed, thus quiting the "look around" state, deleting `this.stopSurveying`, which was later `++`-ed.
That lead to an `NaN` property which got serialised. `NaN` serialisation crashes as of r25151 / rP25151.

Differential revision: https://code.wildfiregames.com/D4177
Reported in: https://wildfiregames.com/forum/topic/41264-alpha-25-pre-releaserelease-candidate-build-testing/page/4/?tab=comments#comment-435709
Reviewed by: @wraitii



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@25804 3db68df2-c116-0410-a063-a993310a9797