Increase the spacing in the top panel and synchronize the colour in the tooltip for gatherers.
With the addition of the new resource icons for stone and metal, the text was too close to the icon so the distance between the text and the icon is increased.
Also, synchronises the colour of the number of gatherers in the top panel with the colour in the tooltip explanation.

Patch by: @Langbart
Differential revision: https://code.wildfiregames.com/D4032
Comments by: @Stan, @wraitii



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@25922 3db68df2-c116-0410-a063-a993310a9797