Use transform for changing a formation template.
As noted on the forums by @Ceres (https://wildfiregames.com/forum/topic/44848-proposals-for-formations/?do=findComment&comment=453840) formations change orientation when changing their template.
This uses the transform helper to change the template of the formation, which apparently also fixes the rotation issue.

With the notion that a different design of transform is preferential (P46).

Differential revision: https://code.wildfiregames.com/D4272
Comments by: @Angen, @bb, @Ceres, @wraitii
Tested by: @Ceres


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@25949 3db68df2-c116-0410-a063-a993310a9797