Remove RENDERDATA_UPDATE_COLOR, which is not used, and cleanup.
RENDERDATA_UPDATE_COLOR was used to precompute lightEnv-dependent data on the CPU. This is no longer done following engine upgrades, and in particular rP24124 which explictly always did this on the GPU.

ModelAbstract had a 'SetDirtyRec' hack for it because of decals, which can also be removed. The 'dirty' bit of CRenderableObject is renderdata for the specific item, never its props, so it never actually needs to be recursive.

CheckLightEnv is also useless as a result, and removed.

Differential Revision: https://code.wildfiregames.com/D4453



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@26249 3db68df2-c116-0410-a063-a993310a9797