Call SetTerrainDirty on CModelAbstract without going through CmpVisualActor
The UnitManager already lists all units, so we do not need to go through the visual actor of entities to update them. This is faster and decouples simulation & graphics code slightly.

Further, the simulation does not need to know about texture changes (see also r26269), so remove those calls in Atlas.

Differential Revision: https://code.wildfiregames.com/D4455



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@26270 3db68df2-c116-0410-a063-a993310a9797