Run the AI in the same Compartment as the simulation. Let the AI access Sim data.
This is a paradigm change for AI computation.
Historically, the AI was intended to be run in a separate thread from the simulation. The idea was that slow AI wouldn't stop the renderer from being smooth.

In that original design, the AI received a copy of the game world and used that to run its logic. This meant the simulation could safely do whatever it wanted in the meantime. This copy was done via AIProxy & AIInterface.

This design ended up having significant flaws:
- The copying impacts the simulation negatively, particularly because AIProxy subscribes to a lot of messages (sometimes sent exclusively to it). This time cannot be threaded, and impacts MP games without AIs.
- Copying the data is increasingly difficult. Modifiers are a headache, LOS is not implemented. Lots of logic is duplicated.

The intended benefits of the design also failed to realise somewhat:
- The AI was never threaded, and in fact, it is probably better to try and thread Sim + AI from the renderer than just the AI, at which point threading the AI specifically brings little benefit.

The new design is much simpler and straighforward, but this has some side-effects:
- The AI can now change the simulation. This can be used for cheating, or possibly for a tutorial AI.
- The AI runs in the same GC zone as the simulation, which may lead to more frequent Sim GCs (but overall we might expect a reduction in temporary objects).
- The AI state was essentially cached, so replacing some functions with Engine.QueryInterface might be slower. The tradeoff should be balanced by lower AIProxy computation times.

Future work:
- Threading some specific AI tasks could still be worthwhile, but should be done in specific worker threads, allowed to run over several turns if needed.

Technical note: the AI 'global' is in its own Realm, which means name collisions with the same are not possible.

Other notes:
- The RL Interface uses the AI Interface and thus will gradually lose some data there. Given that the RL Interface can now request data however, this should be dine.

Refs #5962, #2370

Differential Revision: https://code.wildfiregames.com/D3769



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@26274 3db68df2-c116-0410-a063-a993310a9797