Replace `chicken` in the random maps folder with `startingAnimal`.
Which is more generic. It means we don't have `chicken = oPeacock` and the like anymore.
It adds a default scaling function such that people replacing the animals are puzzled by the different number of starting animals./To keep the amount of food roughly equal to the amount standard provide by chicken.

Patch by: @marder
Differential revision: https://code.wildfiregames.com/D4472
Comments by: @smiley, @Stan



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@26381 3db68df2-c116-0410-a063-a993310a9797