Display defeated/won before victory.
When the victory conditions dictate that someone loses when e.g. losing their CC, the message that the other player(s) had won was sent before the former player was actually defeated.
When the victory conditions dictate that someone wins when e.g. capturing all relics, the message that they have won was sent after the message that the rest has lost.

This changes those.

Differential revision: https://code.wildfiregames.com/D4478
Comments by: @bb, @Silier



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@26382 3db68df2-c116-0410-a063-a993310a9797