[Han Chinese - Alpha 26] Champions trainable in captured Barracks and Stables now require Unlocking Techs
Discussion: https://wildfiregames.com/forum/topic/71582-fixing-chinese-han/

Previously, one could capture an enemy Barracks or Stable and train Han champions there without researching the champion unlocking techs. This rectifies this (in an admittedly hacky way) by creating duplicate Han champion templates with suffix "_academy" and putting those in the Han Academy with the RequiredTechnology of City Phase.

The original Han champion templates, which are what can be trained in captured Barracks and Stables, are given the unlocking RequiredTechnology respective to their class (as most of civs' champions).

Both Templates have the same SelectionGroupName, so should be selectable together with a double-click with no problem.

Please report problems in the above linked thread. 

git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@26632 3db68df2-c116-0410-a063-a993310a9797