Fix non-visual autostart, as well as client/host autostart.
Follows rP26584.

Piping the autostart through a GUI page is easy but requires using the GUI in non-visual mode, which is problematic since it's not initialized.
The GUI is not needed, only its ability to load scripts, so this diff fixes the problem by adding an 'entrypoint' script in the new `autostart` folder that gets called with this unique ability, setting things up properly. The mod mod gets a placeholder.

Other changes:
- Fix some issues with networked autostart:
  - Players were not assigned slots.
  - The host didn't actually wait for other players.
- Move gamesettings from `gui/gamesettings/` to `gamesettings/`. This moves attributes as well.
- Move autostart files from `gui/autostart/` to ` autostart/`. Note that `gui/autostart/` still exists as it's used in the tutorial button from the main menu.
- Rename the JSI_VFS functions to be explicitly about their function: the ability to write and the potential restricted folders.
- Don't require `autostart` when using `--autostart-client` in non-visual mode.

Starting the autostart client/host has poor UX in some situations, but improving that feature is beyond the scope of this diff.

Suggestions for future work:
- Some files in gui/common should probably be moved elsewhere
- Template loading code is duplicated
- The gamesetup code still needs work for proper UI/logic separation.

Help & comments by: smiley

Reported by: vladislav / Langbart

Fixes #6513

Differential Revision: https://code.wildfiregames.com/D4628



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@26879 3db68df2-c116-0410-a063-a993310a9797