Fix text alignment handling of spaces around wrapping.
Follows rP26889.

There is an issue with text-wrapping and word separators (aka spaces).
Because 0 A.D. collates the space after a word to the same TextCall, we occasionally need to ignore it when considering line-wrapping, because we don't want empty spaces on the right-side of right-aligned text.
However, the logic to handle this is currently broken and inconsistent.

The method introduced here uses the SFeedback structure to properly report it and generalises the checks.

Note that multiples spaces are not collapsed in 0 A.D., and for consistency the word-separator-collapsing behaviour is ignored.

Comments by: phosit, vlasdislavbelov

Fixes #6551

Differential Revision: https://code.wildfiregames.com/D4662



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@26915 3db68df2-c116-0410-a063-a993310a9797