[Gameplay] Han unit roster restructure and other balancing adjustments
After feedback on the first RC's : https://wildfiregames.com/forum/topic/83323-balancing-the-han/ 
This version of the roster seems to have a minimal concensus.
This removes some units from the very big roster of the Han, restructures it and changes the cavalry crossbow champion for an infantry one due to histrorical issues.
It furthermore reverts the changes to the champion infantry crossbow template introduced in rP26469 as that nerf wasn't necesarry.
This means that the Han and the Macedonians now have a champion crossbow unit with the same stats.

Also:

Fixes #6560 by only keeping the "academy" templates, so that the Han can now only train champions from the academy. Therefore, this reverts parts of rP26632 and all of rP26633.

Discussion with : @borg- @real_tabasco_sauce @wowgetoffyourcellphone AIEND @Stan and other people on the forums and the Diff.

Differential revision: https://code.wildfiregames.com/D4713

git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@27009 3db68df2-c116-0410-a063-a993310a9797