Fix victory conditions loading (fixes Hero Garrison in regicide crashloop).
VictoryConditions.js gamesetting sends multiple updates to 'active' and 'disabled', instead of just one single update at the end. This is listened to by amongst others the regicide 'hero garrison' setting, thus `HeroGarrison` is deactivated, and then tries to be reactivated from the init attributes, and then disabled again later.

This implements a simple fix, VictoryConditions only sends a single update in FromInitAttributes with the end-state.

Differential Revision: https://code.wildfiregames.com/D4793



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@27135 3db68df2-c116-0410-a063-a993310a9797