Fix and cleanup the SM91 build, fixes #6676.
Add a new patch written by s0600204, replacing the use of `distutils` with `sysconfig` in the setup of the bundled virtualenv in SpiderMonkey.
Add a patch which is a portion of https://phabricator.services.mozilla.com/D130410, fixing upstream's bug 1739486 on Fedora.

This fixes build issues on recent Debian and Fedora. This is consistent with upstream's upcoming changes in ESR102 following the deprecation of `distutils`.

This cleans up some old SM patches; renames the r27442 unicode patch; moves a Python-related environment variable inside `build.sh`; fixes indentation and EOL properties.

Differential Revision: https://code.wildfiregames.com/D4893

git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@27460 3db68df2-c116-0410-a063-a993310a9797