Use premake5 beta2's `externalincludedirs` when available
As of the `beta2` version of `premake5`, `sysincludedirs` has been deprecated
and replaced with/renamed to `externalincludedirs`, and continuing to use it
causes warnings to be emitted.

With this change, we now use `externalincludedirs` when available so as to
prevent the warnings, falling back to `sysincludedirs` when it's not to prevent
breakage with pre-`beta2` versions of `premake5`.


Accepted By: sera
Fixes: #6785
Differential Revision: https://code.wildfiregames.com/D4980



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@27606 3db68df2-c116-0410-a063-a993310a9797