Move some engine required files to the mod mod.
Most of the shaders are explicitely interned by the engine except for the one called "Model" which is not required (you can have maps with just terrain for visualization)
The rng files are called by the engine to validate structure.
The game has two overrides for high quality and normal water to be "ocean" and "default" respectively
The minimap flare folder is hardcoded see #6795
The default material now calls the dummy shader instead of the model one when in modmod
Move the default skybox, it will be required by _default.xml 
The terrain materials are moved as well, for completeness, although they are currently not referenced.

This commit does not include atlas needed files as it cannot run without the public mod anyway for now. It will be done in a separate commit when we have the ingame editor that will require _default.xml for instance.

This commit is also needed to generate the spir-v shaders in the correct mods, in order to be able to launch the game with mod mod only.

Refs: #6636 #5366
Fixes: #6294

Differential Revision: https://code.wildfiregames.com/D4906


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@27629 3db68df2-c116-0410-a063-a993310a9797