Don't poll territory in Position component
CmpPosition::TurnStart checks whether the territory changed underneath each entityevery turn. The only user of this is TerritoryDecay (structures, for the most part). It is rather inefficient to have this done for all entities.
The simplest solution is to listen to position-changed messages in TerritoryDecay instead. This should hardly ever happen in vanilla 0 A.D., except in Atlas, so it's basically free.

This sort of reverts rP14754 (original implementation) and rP14768.

Accepted By: Freagarach (concept only)

Differential Revision: https://code.wildfiregames.com/D5009



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@27673 3db68df2-c116-0410-a063-a993310a9797