Speed up timer update by using Map.forEach
According to JIT profiling on SM 115, using ForEach is faster than iterating.
This is a micro optimisation, maybe 2% faster for OnUpdate, which means we might see an overall improvement < 0.4%.

The reason this is faster here is that Iterating needs to construct an array to store the result, whereas ForEach just uses the values directly. This means this probably doesn't apply to iterating if we don't need both the key/value pair.

Differential Revision: https://code.wildfiregames.com/D5010



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@27674 3db68df2-c116-0410-a063-a993310a9797