Turn off SPECTRE mitigation in jit code
SPECTRE attacks mitigations were added to SpiderMonkey's JIT output in version 57. Turning these off is a very large speedup, around 10-20% wall time (on my computer) depending on the map and the situation.
For the most part, from profiling on my machine, this is from memory fences after JIT -> C++ calls. 0 A.D. does a lot of these.

The SPECTRE class of attack is a timing attack based on speculative execution to leak sensitive information, and it seems extraordinarily unlikely that something like this could be successfully mounted using 0 A.D.

Differential Revision: https://code.wildfiregames.com/D5014



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@27699 3db68df2-c116-0410-a063-a993310a9797