Fix 'Options' asking for confirmation without changes during a game.
Since rP27192, opening the 'Options' menu in-game will show that there are pending changes, even if the user has done nothing.
Unfortunately the cinematic logic to hide silhouettes, which comes from rP19797, is broken following those changes. We use the configuration to hide them, which results in us, indeed, changing the configuration.

There isn't really a good short term fix. This change only attemps to reset the setting if we changed it for a cinematic, which fixes the issue, but has drawbacks.
Ideally, we would use a superior config level, but we need to be careful about exposing that to JS.

Reported by: langbart

Fixes #6821

Differential Revision: https://code.wildfiregames.com/D5040



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@27730 3db68df2-c116-0410-a063-a993310a9797