Bring back (small) splash for catapults, heavy ship.
Catapults and heavy warships received a nerf to their range, which was also designed to stop infinite targeting of structures outside of vision range. This was a substantial nerf, from 100 range to 85. This change compensates for this range reduction by slightly improving their versatility.
This is not the same splash we have seen in previous alphas, as it is only a radius of one meter. It also makes them exciting as you could land some epic shots if your enemy is very tightly bunched.

Patch by: @real_tabasco_sauce
Differential revision: https://code.wildfiregames.com/D5053
Accepted by: @borg-
Comments by: @chrstgtr, @wowgetoffyourcellphone



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@27831 3db68df2-c116-0410-a063-a993310a9797